Struct brotli2::read::BrotliEncoder

pub struct BrotliEncoder<R: Read> {
    // some fields omitted
}

A compression stream which wraps an uncompressed stream of data. Compressed data will be read from the stream.

Methods

impl<R: Read> BrotliEncoder<R>

fn new(r: R, level: u32) -> BrotliEncoder<R>

Create a new compression stream which will compress at the given level to read compress output to the give output stream.

The level argument here is typically 0-9 with 6 being a good default.

fn new_params(r: R, params: &CompressParams) -> BrotliEncoder<R>

Creates a new encoder with a custom Stream.

The Stream can be pre-configured for multithreaded encoding, different compression options/tuning, etc.

fn get_ref(&self) -> &R

Acquires a reference to the underlying stream

fn get_mut(&mut self) -> &mut R

Acquires a mutable reference to the underlying stream

Note that mutation of the stream may result in surprising results if this encoder is continued to be used.

fn into_inner(self) -> R

Unwrap the underlying writer, finishing the compression stream.