Module brotli2::stream

In-memory compression/decompression streams

Structs

Compress
In-memory state for compressing/encoding data with brotli

CompressParams
Parameters passed to various compression routines.

Decompress
In-memory state for decompressing brotli-encoded data.

Error
Error that can happen from decompressing or compressing a brotli stream.

Enums

CompressMode
Possible choices for modes of compression

Status
Possible status results returned from compressing or decompressing.

Functions

compress_buf
Compresses the data in input into output.

compress_vec
Compresses the data in input into output.

decompress_buf
Decompress data in one go in memory.

decompressed_size
Returns the decompressed size of the given encoded stream.