Struct brotli2::stream::Compress

pub struct Compress {
    // some fields omitted
}

In-memory state for compressing/encoding data with brotli

This stream is at the heart of the I/O encoders and is used to compress data.

Methods

impl Compress

fn new(params: &CompressParams) -> Compress

Creates a new compressor ready to encode data into brotli

fn input_block_size(&self) -> usize

Returns the maximum amount of data that can be internally buffered to get processed at once.

Data is fed into this compressor via the copy_input function, and then it's later compressed via the write_brotli_data function.

fn copy_input(&mut self, input: &[u8])

Feeds data into this compressor.

This compressor can store up to self.input_block_size() bytes internally after which the compress call must be made to generate all output of the compressor.

If too much data is copied in then the next call to compress will generate an error.

fn compress(&mut self, last: bool, force_flush: bool) -> Result<&[u8], Error>

Compresses the internal data in this compressor, returning the output buffer of the compressed data.

After data has been fed to this compressor via the copy_input method, the data is then compressed by calling this method. The last flag indicates whether this is the last block of the input (it should only get passed on EOF), and the force_flush flag indicates whether a new meta-block should be created to flush the internal data.

Returns an error, if any, and otherwise the internal buffer which contains the output data of compressed information.