Struct brotli2::stream::Decompress

pub struct Decompress {
    // some fields omitted
}

In-memory state for decompressing brotli-encoded data.

This stream is at the heart of the I/O streams and is used to decompress an incoming brotli stream.

Methods

impl Decompress

fn new() -> Decompress

Creates a new brotli decompression/decoding stream ready to receive data.

fn decompress(&mut self, input: &mut &[u8], output: &mut &mut [u8]) -> Result<Status, Error>

Decompress a block of input data into a block of output data.

This function will decompress the data in input and place the output in output, returning the result. Possible statuses that can be returned are that the stream is finished, more input is needed, or more output space is needed.

The input slice is updated to point to the remaining data that was not consumed, and the output slice is updated to point to the portion of the output slice that still needs to be filled in.

Errors

If the input stream is not a valid brotli stream, then an error is returned.

fn decompress_vec(&mut self, input: &mut &[u8], output: &mut Vec<u8>) -> Result<Status, Error>

Decompress a block of input data into the remaining capacity of a vector.

This function is the same as decompress except that it will fill up the remaining capacity in a destination vector and update the length as necessary.