Struct brotli2::write::BrotliEncoder

pub struct BrotliEncoder<W: Write> {
    // some fields omitted
}

A compression stream which will have uncompressed data written to it and will write compressed data to an output stream.

Methods

impl<W: Write> BrotliEncoder<W>

fn new(obj: W, level: u32) -> BrotliEncoder<W>

Create a new compression stream which will compress at the given level to write compress output to the give output stream.

fn new_params(obj: W, params: &CompressParams) -> BrotliEncoder<W>

Creates a new encoder with a custom CompressParams.

fn finish(self) -> Result<W>

Consumes this encoder, flushing the output stream.

This will flush the underlying data stream and then return the contained writer if the flush succeeded.